new class CreateBook with create_book method
