+++ lib/classes/book.rb
