+++ lib/classes/classroom.rb
