+++ lib/classes/corrector.rb
