+++ lib/classes/person.rb
