+++ lib/classes/rental.rb
