+++ lib/classes/student.rb
