+++ lib/classes/teacher.rb
